requirements.txt
python-docx
openai
pinecone-client
python-dotenv